Move test case into DataWriterAvroTest
